Stock_Project/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
yfinance==0.2.43
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
prophet==1.1.6
scikit-learn==1.5.2
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.8.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
ta==0.11.0
feedparser==6.0.11
